Combine view schedule and edit page, move manager features from view to edit
