+++ client/src/App.tsx
@@ -8,7 +8,6 @@ import Home from "@/Home.tsx";
 import Create from "@/Create.tsx";
 import Profile from "@/Profile.tsx";
 //import EditProfile from "@/EditProfile.tsx"; // To be added later
-import Schedule from "@/schedule/Schedule.tsx";
 import EditTree from "@/schedule/EditTree";
 import NotFoundPage from "@/NotFound.tsx";
 
@@ -16,6 +15,7 @@ import { ThemeProvider } from "@mui/material/styles";
 import customTheme from "./theme";
 import { NotifierProvider } from "@/notifier.tsx";
 import SignUpConfirmation from "./SignUpConfirmation.tsx";
+import ScheduleShared from "@/schedule/ScheduleShared.tsx";
 
 export default function App() {
   return (
@@ -79,7 +79,7 @@ export default function App() {
                   path="/schedule/:scheduleId"
                   element={
                     <Authenticated>
-                      <Schedule />
+                      <ScheduleShared />
                     </Authenticated>
                   }
                 />
