+++ client/src/customComponents/CustomTooltip.tsx
@@ -1,0 +1,17 @@
+import {
+  styled,
+  Tooltip,
+  tooltipClasses,
+  type TooltipProps,
+} from "@mui/material";
+
+export const CustomTooltip = styled(({ className, ...props }: TooltipProps) => (
+  <Tooltip {...props} arrow classes={{ popper: className }} />
+))(({ theme }) => ({
+  [`& .${tooltipClasses.arrow}`]: {
+    color: theme.palette.common.black,
+  },
+  [`& .${tooltipClasses.tooltip}`]: {
+    backgroundColor: theme.palette.common.black,
+  },
+}));
