+++ client/src/schedule/Schedule.tsx
@@ -8,11 +8,6 @@ import {
   Chip,
   Box,
   Slider,
-  Avatar,
-  Tooltip,
-  TooltipProps,
-  tooltipClasses,
-  styled,
 } from "@mui/material";
 import { useParams } from "react-router";
 import {
@@ -29,22 +24,10 @@ import Navbar from "@/Navbar";
 import NavbarPadding from "@/NavbarPadding";
 import EditShiftDrawer from "./EditShiftDrawer";
 import { ShiftCalendar, ShiftDetails } from "./ShiftCalendar";
-import { createRandomPfpUrl } from "./EditMembersTab";
 import { useEmployeeActions } from "@/hooks/useEmployeeActions";
 import theme from "@/theme";
 import { useNotifier } from "@/notifier";
-import { useShifts } from "@/hooks/useShifts";
-
-const CustomTooltip = styled(({ className, ...props }: TooltipProps) => (
-  <Tooltip {...props} arrow classes={{ popper: className }} />
-))(({ theme }) => ({
-  [`& .${tooltipClasses.arrow}`]: {
-    color: theme.palette.common.black,
-  },
-  [`& .${tooltipClasses.tooltip}`]: {
-    backgroundColor: theme.palette.common.black,
-  },
-}));
+import { CustomTooltip } from "@/customComponents/CustomTooltip";
 
 function useSelectedShiftParam() {
   const [searchParams, setSearchParams] = useSearchParams();
@@ -140,26 +123,6 @@ export default function Schedule() {
     () => selectedShift && signedUpShifts.includes(selectedShift),
     [selectedShift, signedUpShifts],
   );
-
-  function ManagerPerShiftStackContent(props: { shiftIds: string[] }) {
-    const shiftIds = useMemo(() => new Set(props.shiftIds), [props.shiftIds]);
-    const assignedUsers = useMemo(
-      () =>
-        empActions.allAssignments
-          ?.filter(asgn => asgn.shiftId && shiftIds.has(asgn.shiftId))
-          .map(asgn => asgn.user)
-          .filter(u => u !== undefined) ?? [],
-      [shiftIds],
-    );
-
-    return (
-      <>
-        {assignedUsers.map(user => (
-          <UserIndicators key={user.id} name={user.displayName} id={user.id} />
-        ))}
-      </>
-    );
-  }
 
   function MemberPerShiftStackContent(props: { shiftIds: string[] }) {
     const isRegistered = useMemo(
@@ -288,17 +251,6 @@ export default function Schedule() {
                 </Box>
               </>
             )}
-            {isManager && (
-              <>
-                <Typography variant="h6">Registered Members</Typography>
-                {/* Chips for users that are signed up */}
-
-                <UserChips
-                  scheduleId={scheduleId}
-                  shiftId={selectedShift ?? undefined}
-                ></UserChips>
-              </>
-            )}
           </EditShiftDrawer>
           <ShiftCalendar
             onClickShift={shiftId => setSelectedShift(shiftId)}
@@ -306,11 +258,7 @@ export default function Schedule() {
             endDate={dayjs(scheduleData?.endTime ?? dayjs().toISOString())}
             selectedShifts={selectedShift ? [selectedShift] : []}
             shifts={formattedShifts}
-            CustomContent={
-              isManager
-                ? ManagerPerShiftStackContent
-                : MemberPerShiftStackContent
-            }
+            CustomContent={MemberPerShiftStackContent}
           />
         </Paper>
       </Container>
@@ -342,69 +290,5 @@ function AssignedIndicator() {
       label="Assigned"
       color="primary"
     />
-  );
-}
-
-interface UserIndicatorsProps {
-  name: string;
-  id: string;
-}
-
-function UserIndicators(props: UserIndicatorsProps) {
-  return (
-    <Chip
-      avatar={<Avatar src={createRandomPfpUrl(props.name, props.id)}></Avatar>}
-      sx={{
-        backgroundColor: theme.palette.primary.main,
-        color: "white",
-      }}
-      label={props.name}
-      color="primary"
-    />
-  );
-}
-
-interface UserChipsProps {
-  scheduleId?: string;
-  shiftId?: string;
-}
-
-function UserChips(props: UserChipsProps) {
-  const api = useApi();
-
-  // This request is required to get the users that are signed up in each schedule
-  const { data: scheduleSignups } = api.useQuery(
-    "get",
-    "/schedules/{scheduleId}/signups",
-    { params: { path: { scheduleId: props.scheduleId as string } } },
-  );
-
-  const shifts = useShifts(props.scheduleId ?? "");
-
-  const stackShiftIds = useMemo(
-    () => new Set(shifts.matchingShifts(props.shiftId ?? "").map(s => s.id)),
-    [props.shiftId, shifts],
-  );
-
-  const users = scheduleSignups
-    ?.filter(shift => stackShiftIds.has(shift.id)) // Match the shiftId
-    .flatMap(shift => shift.signups?.map(signup => signup.user))
-    .filter(u => u !== undefined);
-
-  return (
-    <Box>
-      {users?.map(user => {
-        return (
-          <Chip
-            key={user.id}
-            avatar={
-              <Avatar src={createRandomPfpUrl(user.displayName, user.id)} />
-            }
-            label={user.displayName}
-            variant="outlined"
-          />
-        );
-      })}
-    </Box>
   );
 }
